subdivision de la classe camerachess
